bookstore-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy>=2.0.36
pymysql==1.1.0
cryptography==41.0.7
alembic==1.12.1
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
fastapi-mail==1.4.1
python-dotenv==1.0.0
# Pillow 10.1.0 fails to build on Python 3.13; use 11+ which provides wheels
pillow>=11.0.0
aiofiles==23.2.1
pydantic[email]>=2.7.0
httpx==0.27.2
pytest==7.4.3
pytest-asyncio==0.21.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
chromadb==0.5.10
sentence-transformers==2.7.0
groq==0.12.0
bcrypt==4.0.1